pyproject.toml
[tool.setuptools]
packages = ["tft_info_collector"]

[project]
name = "tft-info-collector"
version = "0.2.0"
description = "Unofficial TFT data collection and analysis CLI"
readme = "README.md"
requires-python = ">=3.9"
license = "GPL-3.0-or-later"

dependencies = [
    "httpx",
    "requests",
    "ujson",
    "pydantic-settings",
    "python-dotenv",
    "typer",
    "rich",
]

[project.scripts]
tft-collector = "tft_info_collector.cli:main"